Split admin from people
