+++ server/lib/admin/stream.js
